+++ test/fixtures/my-lib.js
@@ -9,6 +9,12 @@ exports.old2 = function () {
   deprecate('old2')
 }
 
+exports.oldfn = deprecate.function(fn, 'oldfn')
+
+exports.oldfnauto = deprecate.function(fn)
+
+exports.oldfnautoanon = deprecate.function(function () {})
+
 exports.automsg = function () {
   deprecate()
 }
@@ -19,4 +25,8 @@ exports.automsgnamed = function automsgnamed() {
 
 exports.automsganon = function () {
   (function () { deprecate() }())
+}
+
+function fn(a1, a2) {
+  return a2
 }